Ready for testing?

+++ Core/Inc/channels/servo_channel.h
@@ -5,8 +5,8 @@
 
 typedef struct
 {
-	uint_least16_t position_fb;
-	uint_least16_t position_set;
+	uint32_t position_fb;
+	uint32_t position_set;
 	//other servo stuff
 }Servo_Channel_t;
